requirements.txt
asgiref==3.6.0
Django==4.1.6
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.3.0
mongoengine==0.26.0
PyJWT==1.7.1
pymongo==3.12.1
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.2.4
uuid==1.30
